pyproject.toml
[project]
name = "reservas-api"
version = "0.1.0"
description = "Meeting Room Reservation API with FastAPI, PostgreSQL and SQLAlchemy"
authors = [
    {name = "", email = ""},
]
dependencies = ["fastapi>=0.121.0", "uvicorn>=0.38.0", "sqlalchemy>=2.0.44", "psycopg2-binary>=2.9.11", "python-dotenv>=1.2.1", "pydantic>=2.12.3", "fastapi-cors>=0.0.6", "google-auth>=2.23.0", "pyjwt>=2.8.0", "cryptography>=41.0.0", "requests>=2.32.5", "alembic>=1.13.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
